fix all the basic stuff i did wrong the first time, renamed all classes and folders
