+++ gradle.properties
@@ -6,9 +6,9 @@ yarn_mappings=1.21+build.8
 loader_version=0.15.11
 
 # Mod Properties
-mod_version=0.1.0
-maven_group=com.example
-archives_base_name=addon-template
+mod_version=0.2.0
+maven_group=com.ivanprocookie
+archives_base_name=ModulesPlus
 
 # Dependencies
