+++ replit.nix
@@ -1,0 +1,5 @@
+{pkgs}: {
+  deps = [
+    pkgs.jdk21
+  ];
+}
