+++ src/main/java/com/ivanprocookie/addon/mixin/ExampleMixin.java
@@ -1,6 +1,6 @@
-package com.example.addon.mixin;
+package com.ivanprocookie.addon.mixin;
 
-import com.example.addon.AddonTemplate;
+import com.ivanprocookie.addon.ModulesPlus;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.RunArgs;
 import org.spongepowered.asm.mixin.Mixin;
@@ -26,6 +26,6 @@ public abstract class ExampleMixin {
      */
     @Inject(method = "<init>", at = @At("TAIL"))
     private void onGameLoaded(RunArgs args, CallbackInfo ci) {
-        AddonTemplate.LOG.info("Hello from ExampleMixin!");
+        ModulesPlus.LOG.info("Hello from ExampleMixin!");
     }
 }
