+++ src/main/resources/modules-plus.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "com.example.addon.mixin",
+  "package": "com.ivanprocookie.addon.mixin",
   "compatibilityLevel": "JAVA_21",
   "client": [
     "ExampleMixin"
